% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingoverlap-package.R
\docType{package}
\name{fishingoverlap}
\alias{fishingoverlap}
\alias{fishingoverlap-package}
\title{fishingoverlap: A package for estimating the spatial overlap between
fishing events}
\description{
fishingoverlap: A package for estimating the spatial overlap between
fishing events
}
